🎨 Added collectAndSaveImg.cs
scriptOutput.cs collectAndSaveImg has moved to collectAndSaveImg.cs
