+++ On The Land/Assets/Resources/Scripts/CollectAndSaveImg.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a108de73b192a6541b9a8455dfad334c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
