+++ On The Land/Assets/Resources/Scripts/autoScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 24eb8dc0cc3a9a74388dcf75f3f85985
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
