+++ On The Land/UserSettings/EditorUserSettings.asset
@@ -24,16 +24,16 @@ EditorUserSettings:
       value: 050901515c060f025e5c0f7745760d444616192e2f2b71607e2b4c61bbb96d69
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
-      flags: 0
-    RecentlyUsedSceneGuid-7:
       value: 5a0956050751505d5457097a44730844134e1d2f2f2b71347f7f1e30e4b6613d
       flags: 0
-    RecentlyUsedSceneGuid-8:
+    RecentlyUsedSceneGuid-7:
       value: 540256575c040f025b0b0a2443760844144e4173792c77602b2f4b62e7e26d6e
       flags: 0
-    RecentlyUsedSceneGuid-9:
+    RecentlyUsedSceneGuid-8:
       value: 5501075707005f0d090f0d21147a0844444e4979297d746774791966b6b6316b
+      flags: 0
+    RecentlyUsedSceneGuid-9:
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
